7.10 新增了连接验证、超时服务，将图表以<views:的形式嵌入到对话框，修改了仓储层读取数据的逻辑，先读数据，再在内存中操作。注册接口到各项目的配置文件中。

+++ IndustrialMonitor.Core/GraphModels/GraphDataModel.cs
@@ -10,6 +10,6 @@ namespace IndustrialMonitor.Core.GraphModels
     {
         public DateTime dateTime;
 
-        public ushort value;
+        public double value;
     }
 }
